requirements.txt
discord.py
EorzeaEnv
pytz